Delete Code of Conduct
This overly political and subjective Code of Conduct was created without the knowledge or consent of contributors and maintainers. I certainly do not pledge to arbitrarily police conduct in conflict with Post-modernist critical theory. This is an open forum for anyone, and while I don't foresee any insults being thrown around, we (devSparkle and I) already have the power to remove any comments and remove anyone from our repositories. The answer is absolutely not.